Address some of PR comments
