+++ support/sparse_mmap/src/unix.rs
@@ -95,14 +95,6 @@ impl SparseMapping {
     /// The range will be aligned to the largest system page size that's smaller
     /// or equal to `len`.
     pub fn new(len: usize) -> Result<Self, Error> {
-        Self::new_at(len, None)
-    }
-
-    /// Reserves a sparse mapping range with the given size at specific GPA.
-    ///
-    /// The range will be aligned to the largest system page size that's smaller
-    /// or equal to `len`.
-    pub fn new_at(len: usize, fixed_addr: Option<u64>) -> Result<Self, Error> {
         super::initialize_try_copy();
 
         // Length of 0 return an OS error, so we need to handle it explicitly.
@@ -144,11 +136,10 @@ impl SparseMapping {
                 )
             })?;
 
-        let libc_addr = fixed_addr.map_or(null_mut(), |a| a as *mut c_void);
         // SAFETY: calling mmap to allocate a new range.
         let address = unsafe {
             mmap(
-                libc_addr,
+                null_mut(),
                 alloc_len,
                 libc::PROT_NONE,
                 libc::MAP_PRIVATE | libc::MAP_ANONYMOUS,
@@ -233,7 +224,7 @@ impl SparseMapping {
             libc::PROT_READ
         };
 
-        // SAFETY: The flags passed in are guaranteed to be valid. MAP_SHARED is required.
+        // SAFETY: The flags passed in are guaranteed to be valid
         unsafe {
             self.mmap(
                 offset,
