+++ vm/devices/storage/disk_nvme/nvme_driver/src/lib.rs
@@ -16,7 +16,7 @@ mod registers;
 mod tests;
 
 pub use self::driver::NvmeDriver;
-pub use self::driver::NvmeDriverSavedState;
+pub use self::driver::save_restore::NvmeDriverSavedState;
 pub use self::namespace::Namespace;
 pub use self::namespace::NamespaceError;
 pub use self::queue_pair::RequestError;
