+++ vm/devices/storage/disk_nvme/nvme_driver/src/queues.rs
@@ -4,9 +4,10 @@
 //! Implementation of submission and completion queues.
 
 use super::spec;
+use crate::driver::save_restore::CompletionQueueSavedState;
+use crate::driver::save_restore::SubmissionQueueSavedState;
 use crate::registers::DeviceRegisters;
 use inspect::Inspect;
-use mesh::payload::Protobuf;
 use user_driver::memory::MemoryBlock;
 use user_driver::DeviceBacking;
 
@@ -77,7 +78,6 @@ impl SubmissionQueue {
             tail: self.tail,
             committed_tail: self.committed_tail,
             len: self.len,
-            base_mem: self.mem.base_va(),
             pfns: self.mem.pfns().to_vec(),
         }
     }
@@ -152,7 +152,6 @@ impl CompletionQueue {
             committed_head: self.committed_head,
             len: self.len,
             phase: self.phase,
-            base_mem: self.mem.base_va(),
             pfns: self.mem.pfns().to_vec(),
         }
     }
@@ -174,43 +173,4 @@ fn advance(n: u32, l: u32) -> u32 {
     } else {
         0
     }
-}
-
-#[derive(Protobuf, Clone, Debug)]
-#[mesh(package = "underhill")]
-pub struct SubmissionQueueSavedState {
-    #[mesh(1)]
-    pub sqid: u16,
-    #[mesh(2)]
-    pub head: u32,
-    #[mesh(3)]
-    pub tail: u32,
-    #[mesh(4)]
-    pub committed_tail: u32,
-    #[mesh(5)]
-    pub len: u32,
-    #[mesh(7)]
-    pub base_mem: u64,
-    #[mesh(8)]
-    pub pfns: Vec<u64>,
-}
-
-#[derive(Protobuf, Clone, Debug)]
-#[mesh(package = "underhill")]
-pub struct CompletionQueueSavedState {
-    #[mesh(1)]
-    pub cqid: u16,
-    #[mesh(2)]
-    pub head: u32,
-    #[mesh(3)]
-    pub committed_head: u32,
-    #[mesh(4)]
-    pub len: u32,
-    #[mesh(5)]
-    /// NVMe completion tag.
-    pub phase: bool,
-    #[mesh(7)]
-    pub base_mem: u64,
-    #[mesh(8)]
-    pub pfns: Vec<u64>,
 }
